delete account to facade
